trnasformations general structure
